refactor update button into it's own view
