+++ js/templates/modals/settings/bulkCoinUpdateBtn.html
@@ -1,0 +1,18 @@
+<%= ob.processingButton({
+className: `btn clrP clrBr clrT clrSh2 ${ob.isBulkCoinUpdating ? 'processing' : ''} js-applyToCurrent`,
+btnText: ob.polyT('settings.storeTab.bulkListingCoinUpdate.MainButton'),
+}) %>
+<% if (ob.showBulkConfirm) {%>
+<div class="confirmBox arrowBoxTop clrBr clrP clrT clrSh1 js-confirmBox">
+  <div class="tx3 txB rowSm"><%= ob.polyT('settings.storeTab.bulkListingCoinUpdate.ConfirmTitle') %></div>
+  <div class="posR padSm"><%= ob.polyT('settings.storeTab.bulkListingCoinUpdate.ConfirmMessage') %></div>
+  <hr class="clrBr row" />
+  <div class="flexVCent gutterHLg">
+    <div class="flexExpand">
+      <% // The cancel button is just costmetic, cancelling is handled by the click on the document %>
+      <a class="btnTxtOnly barBtn clrT2 txUnb js-applyToCurrentCancel"><%= ob.polyT('settings.storeTab.bulkListingCoinUpdate.Cancel') %></a>
+    </div>
+    <a class="btn clrBAttGrad clrBrDec1 clrTOnEmph js-applyToCurrentConfirm"><%= ob.polyT('settings.storeTab.bulkListingCoinUpdate.Apply') %></a>
+  </div>
+</div>
+<% } %>
