+++ js/templates/modals/settings/store.html
@@ -53,26 +53,7 @@
             <% if (ob.errors.currencies) print(ob.formErrorTmpl({ errors: ob.errors.currencies })) %>
             <div class="row js-currencySelector"></div>
             <div class="flexHRight">
-              <div>
-              <%= ob.processingButton({
-              className: `btn clrP clrBr clrT clrSh2 ${ob.isBulkCoinUpdating ? 'processing' : ''} js-applyToCurrent`,
-              btnText: ob.polyT('settings.storeTab.bulkListingCoinUpdate.MainButton'),
-              }) %>
-              <% if (ob.showBulkConfirm) {%>
-                <div class="confirmBox arrowBoxTop clrBr clrP clrT clrSh1">
-                  <div class="tx3 txB rowSm"><%= ob.polyT('settings.storeTab.bulkListingCoinUpdate.ConfirmTitle') %></div>
-                  <div class="posR padSm"><%= ob.polyT('settings.storeTab.bulkListingCoinUpdate.ConfirmMessage') %></div>
-                  <hr class="clrBr row" />
-                  <div class="flexVCent gutterHLg">
-                    <div class="flexExpand">
-                      <% // The cancel button is just costmetic, cancelling is handled by the click on the document %>
-                      <a class="btnTxtOnly barBtn clrT2 txUnb"><%= ob.polyT('settings.storeTab.bulkListingCoinUpdate.Cancel') %></a>
-                    </div>
-                    <a class="btn clrBAttGrad clrBrDec1 clrTOnEmph js-applyToCurrentConfirm"><%= ob.polyT('settings.storeTab.bulkListingCoinUpdate.Apply') %></a>
-                  </div>
-                </div>
-              <% } %>
-              </div>
+              <div class="js-bulkCoinUpdateBtn"></div>
             </div>
           </div>
         </div>
